Remove dependencies: jQuery, Bootstrap, fullScreen
This replaces the above three big dependencies with normalize.css and adds a bit of CSS to get the page back to the original look.
The biggest drawback with this solution is that we don't get the nice transitions from fullPage.js, but I think the filesize savings are worth it.
Also, I used a different font stack, which will show native fonts for macOS and Windows.